+++ src/LiteBus.Messaging/Registry/Builders/ErrorHandlerDescriptorBuilder.cs
@@ -21,7 +21,7 @@ internal sealed class ErrorHandlerDescriptorBuilder : IHandlerDescriptorBuilder
     /// <inheritdoc />
     public IEnumerable<IHandlerDescriptor> Build(Type handlerType)
     {
-        var interfaces = handlerType.GetInterfacesEqualTo(typeof(IMessageErrorHandler<,>));
+        var interfaces = handlerType.GetInterfacesEqualTo(typeof(IAsyncMessageErrorHandler<,>));
         var priority = handlerType.GetPriorityFromAttribute();
 
         foreach (var @interface in interfaces)